% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoring.R
\name{spotify_appinfo}
\alias{spotify_appinfo}
\title{Calculate the score of Spotify tracks}
\usage{
spotify_appinfo(genres, features, levels)
}
\arguments{
\item{genres}{A \code{string} contain the name of the genres that the user chooses.}

\item{features}{A \code{list} containing 5 characters. Each character represents
the feature that the user choose}

\item{levels}{A \code{list} containing 5 numerics. Each numeric represent the
value of each feature that user want.}
}
\value{
list_values \code{list} containing the following attributes:
\describe{
     \item{result}{containing the top10 scores tracks}
     \item{user_genre}{contain the genre that user choose}
     \item{user_feature}{contain the features that user choose}
     \item{user_level}{contain the level of each feature that user choose}
}
}
\description{
Calculate the score of Spotify tracks according to the features
and the level for each feature that user choose.
}
\examples{
spotify_appinfo(genres = "Funk",
features=c("Danceability","Energy","Loudness","Speechiness","Liveness"),
levels = c(9,8,5,4,7))
}
\author{
Ting Yang
}
